refactor(services): tmux_service 를 패키지로 승격해 한도 안에 넣는다 (B5.5 Task 2)
`services/tmux_service.py` 920 줄 → 패키지 3 모듈, 최대 618 줄.
`X.py` → `X/__init__.py`(배럴) + 형제 모듈. `from services.tmux_service import
get_tmux_service` 가 글자 그대로 유효하게 남아 프로덕션 호출부(api/agents/tmux.py
6 곳)를 한 줄도 고치지 않는다.

    models.py    30줄  ClaudeAuthStatus · TmuxSessionInfo
    usage.py    294줄  CLI usage 파서 + 원장 기록 브리지
    service.py  618줄  TmuxService + 싱글턴

**모듈 지역 패치 타깃 2 종 — 이 파일의 고유 위험.**
테스트의 문자열 패치 4 건이 전부 `services.tmux_service` 네임스페이스로 가져온
원장 심볼 둘을 겨냥한다(`record_usage_best_effort` 3 건 ·
`enforce_usage_quota_preflight_best_effort` 1 건). 읽는 함수 둘을 `usage.py` 에
모으고 **배럴은 그 둘을 재노출하지 않는다.** 재노출하면 낡은 패치 경로가
성공하되 서브모듈 전역은 원본을 계속 봐서 조용히 무효가 된다. 빼두면
`monkeypatch.setattr` 이 그 줄에서 `AttributeError` 로 죽어 실패가 자기 위치를
가리킨다 — 실측: 재지정 전 3 건이 정확히 그렇게 실패했다(RED), 재지정 후 30 통과.

**싱글턴.** `_tmux_service` 는 `get_tmux_service` 안에서 `global` 로 재바인딩되고
외부 writer 는 0 건이다. 접근자와 같은 모듈에 두고 배럴은 **함수만** 내보낸다 —
값을 재노출하면 Task 1 의 `_audit_logs` 와 같은 스냅샷 버그가 난다.

**검증 (STATE.md 체크리스트 7 단계 전부).**
- `split_audit.py`: 정의 20 종 이름·본문 바이트 보존. 프로브로 Red-Green
  (`_compact_text` 삭제 → FAIL)
- 분할 전후 **실행 비교**: 싱글턴 멱등·모듈 전역 일치, 파서 3 입력, 원장 도달
  경로가 JSON 으로 완전 동일. Task 1 은 이 단계를 건너뛰어 회귀를 만들었다
- 게이트: ruff check · ruff format --check · mypy(320 파일) · pytest 1,698 통과

**부수 수정 2 건.**
1. `scripts/split_audit.py` — `logger` 를 per-module 이름으로 tolerate.
   `split_module.py` 는 이미 제외하는데 audit 쪽만 빠져 있어 "두 모듈에 같은
   정의: logger" 로 중단됐다. Task 1 의 `audit_service` 는 모듈 레벨 logger 가
   없어 드러나지 않았을 뿐이고, 남은 3 개는 전부 같은 곳에서 막힌다.
   유실 자체는 여전히 잡힌다(프로브 FAIL 확인) + ruff F821 이 백스톱(실측)
2. `docs/architecture.md` 트리의 `audit_service.py` — Task 1 이 패키지로 만들고
   갱신하지 않은 잔여 드리프트. 같은 표를 고치는 김에 사실만 정정

엔진의 `_referenced` 가 정규식 과대추정이라 `service.py` 에 불필요한
`import re` 를 얹었다(원본은 메서드 지역 import 를 따로 갖고 있다). audit 이
import 를 의도적으로 대조하지 않고 ruff 에 맡기는 분업이 그 층을 잡았다.
